+++ types/sign.ts
@@ -1,0 +1,5 @@
+export interface SignUpForm {
+  email: string
+  password: string
+  passwordCheck: string
+}
